add texture implementation (OpenGL) add rectangle rendering (single, multiple, colored, textured)

+++ source/engine/graphics/vertex.h
@@ -6,5 +6,5 @@ struct vertex {
     vec3f position{ 0.f };
     color color{ 1.f };
     vec2f tex_coords{ 0.f };
-    int texture_id{ -1 };
+    float texture_id{ -1.f };
 };
